Adjustment for V1.1.0:
* Rename create command to generate
* Refactor Template specific item
* Change yaml structure
* Now it contains a section "creatorly" for the placeholders.
* Refine usage
* Refactor git loader and refine documentation
