+++ .github/workflows/release.yml
@@ -3,7 +3,7 @@ name: Release
 on:
   push:
     tags:
-      - 'V*.*.*'
+      - "V*.*.*"
 
 jobs:
   build-and-release:
