+++ Cargo.toml
@@ -3,7 +3,7 @@ members = ["src/application", "src/infrastructure", "src/cli"]
 
 [workspace.package]
 authors = ["Burak Kizilkaya <burak.kizilkaya@outlook.com>"]
-version = "1.0.4"
+version = "1.1.0"
 
 [workspace.dependencies]
 log = "0.4"
