+++ assets/example_project/creatorly.yml
@@ -1,5 +1,6 @@
-project_name: "DemoBoilerplate"
-description: "Demo Boilerplate"
-author_name: 'Max Muster'
-author_email: 'max.muster@example.com'
-license: ["MIT", "BSD"]
+creatorly:
+  project_name: "DemoBoilerplate"
+  description: "Demo Boilerplate"
+  author_name: 'Max Muster'
+  author_email: 'max.muster@example.com'
+  license: ["MIT", "BSD"]
