+++ assets/example_project/src/index.html
@@ -4,20 +4,20 @@
 <head>
     <meta charset='utf-8'>
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
-    <title>{{ project_name }}</title>
+    <title>{{ creatorly.project_name }}</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
     <link rel='stylesheet' type='text/css' media='screen' href='main.css'>
     <script src='main.js'></script>
 </head>
 
 <body>
-    <h1>{{ project_name }}</h1>
+    <h1>{{ creatorly.project_name }}</h1>
 
-    <p>{{ description }}</p>
+    <p>{{ creatorly.description }}</p>
 
-    <p>From {{ author_name }} ({{ author_email }})</p>
+    <p>From {{ creatorly.author_name }} ({{ creatorly.author_email }})</p>
 
-    <p>Licence {{ license }}</p>
+    <p>Licence {{ creatorly.license }}</p>
 </body>
 
 </html>
