+++ docs/architecture/content/05_building_block_view.adoc
@@ -1,11 +1,7 @@
 ifndef::imagesdir[:imagesdir: ../.images]
 
 [[section-building-block-view]]
-
-
 == Building Block View
-
-
 
 === Whitebox Overall System
