+++ docs/usage/tipps_for_template.adoc
@@ -1,0 +1,16 @@
+= Tipps for Templates
+
+== Exclude text from rendering
+
+If you want to exclude text from rendering, you can use the `{% raw %}blablab{% endraw %}` tag. This is useful for templates that contain a CI pipeline file which s has sombe variables that should not be rendered.
+
+=== Example
+
+[source,yaml]
+----
+
+name: Extract version number
+id: extract_version
+run: echo "{{ creatorly.projectname }}{% raw %}=$(echo ${{ github.ref }} | sed 's/refs\/tags\/V\(.*\)/\1/')" >> $GITHUB_ENV{% endraw %}
+
+----
