+++ src/application/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod create;
 pub mod file_system;
+pub mod generate;
