+++ src/infrastructure/resources/test/test_template_spec.yml
@@ -1,6 +1,7 @@
-project_name: "DemoBoilerplate"
-description: "Demo Boilerplate"
-author_name: 'Max Muster'
-author_email: 'max.muster@example.com'
-license: ["MIT", "BSD"]
+creatorly:
+  project_name: "DemoBoilerplate"
+  description: "Demo Boilerplate"
+  author_name: 'Max Muster'
+  author_email: 'max.muster@example.com'
+  license: ["MIT", "BSD"]
