+++ src/infrastructure/src/folder_loader/local_file_loader.rs
@@ -1,4 +1,4 @@
-use application::create::{file_list::FileList, interfaces::FileListLoader};
+use application::generate::{file_list::FileList, interfaces::FileListLoader};
 use std::path::PathBuf;
 use walkdir::WalkDir;
